#44 service changes for time log
git-svn-id: https://subversion.assembla.com/svn/projectx-0/branches@520 b02b8b71-e8a6-4a5c-b3be-45d727391186

+++ ProjectXService/common/models/mongo/TicketCollection.php
@@ -369,7 +369,7 @@ class TicketCollection extends ActiveRecord
             $query->from('TicketCollection')
                     ->where(['TicketId' => (int) $ticketId, "ProjectId" => $projectId]);
             $ticketDetails = $query->one();
-            $totalTimeLog = (float) ($ticketDetails["TotalTimeLog"] + $totalWorkHours);
+            $totalTimeLog = $ticketDetails["TotalTimeLog"] + $totalWorkHours;
             if ($totalTimeLog > 0) {
                 $ticketCollection = Yii::$app->mongodb->getCollection('TicketCollection');
                 $updateTotalTimeLog = array('$set' => array("TotalTimeLog" => $totalTimeLog));
